refactor(pagination): group pagination classes
