+++ apps/api/src/HttpResource/Pagination/PaginatedResourceFactory.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\HttpResource\Pagination;
 
-use App\HttpResource\HttpPaginatedResource;
 use App\HttpResource\HttpResourceFactory;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
@@ -16,7 +15,7 @@ class PaginatedResourceFactory implements PaginatedResourceFactoryInterface
     ) {
     }
 
-    public function create(PaginatedCollection $collection, string $routeName, array $options): HttpPaginatedResource
+    public function create(PaginatedCollection $collection, string $routeName, array $options): PaginatedResources
     {
         $items = array_map(
             fn($item) => $this->httpResourceFactory->create($item, $options),
@@ -26,7 +25,7 @@ class PaginatedResourceFactory implements PaginatedResourceFactoryInterface
         $currentPage = $collection->page;
         $lastPage = (int)ceil($collection->totalItems / $collection->perPage);
 
-        return new HttpPaginatedResource(
+        return new PaginatedResources(
             items: $items,
             currentPage: $this->urlGenerator->generate($routeName, ['page' => $currentPage]),
             lastPage: $this->urlGenerator->generate($routeName, ['page' => $lastPage]),
