+++ apps/api/src/HttpResource/Pagination/PaginatedResourceFactoryInterface.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\HttpResource\Pagination;
 
-use App\HttpResource\HttpPaginatedResource;
 use Symfony\Component\Serializer\Exception\ExceptionInterface;
 
 interface PaginatedResourceFactoryInterface
@@ -13,5 +12,5 @@ interface PaginatedResourceFactoryInterface
      * @param array{groups: mixed} $options
      * @throws ExceptionInterface
      */
-    public function create(PaginatedCollection $collection, string $routeName, array $options): HttpPaginatedResource;
+    public function create(PaginatedCollection $collection, string $routeName, array $options): PaginatedResources;
 }
